requirements.txt
alembic==1.12.0
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.7
Flask==2.0.2
grpcio
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
Mako==1.2.4
MarkupSafe==2.1.3
protobuf==4.24.3
psycopg2==2.9.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.7
